Improve Discord Bot example (#236)
* Set presence in Discord Bot constructor

It's recommended not to change_presence in an on_ready. Discord has a high chance to completely disconnect you during the READY or GUILD_CREATE events with a 1006 close code if you change_presence during READY and there is nothing you can do to prevent it, so it's better to set the presence in the commands.Bot constructor.

* Use consume rest behaviour in track command

Making the 'query' argument a kwarg in the track command tells discord.py to parse everything instead of just the first word, thus elimating the need for quotes for multi-word searches.